project_1/requirements.txt
pillow
prettier
duckdb
markdown2
fastapi
uvicorn
python-multipart
httpx
python-dotenv
aiofiles
nodejs
python-dateutil
numpy
requests
beautifulsoup4
gitpython
faker
npm
